lib: move arena cpumask from wd40 into main library directory
