+++ lib/cpumask.bpf.c
@@ -1,7 +1,7 @@
 #include <scx/common.bpf.h>
 #include <lib/sdt_task.h>
 
-#include "cpumask.h"
+#include <lib/cpumask.h>
 #include <lib/percpu.h>
 
 static struct sdt_allocator scx_bitmap_allocator;
