+++ scheds/rust/scx_wd40/build.rs
@@ -12,8 +12,8 @@ fn main() {
         .add_source("src/bpf/common.bpf.c")
         .add_source("src/bpf/deadline.bpf.c")
         .add_source("src/bpf/placement.bpf.c")
-        .add_source("src/bpf/cpumask.bpf.c")
         .add_source("src/bpf/deadline.bpf.c")
+        .add_source("../../../lib/cpumask.bpf.c")
         .add_source("../../../lib/sdt_task.bpf.c")
         .add_source("../../../lib/sdt_alloc.bpf.c")
         .compile_link_gen()
