+++ scheds/rust/scx_wd40/src/bpf/common.bpf.c
@@ -11,7 +11,7 @@
 #include <scx/bpf_arena_common.h>
 #include <scx/bpf_arena_spin_lock.h>
 
-#include "cpumask.h"
+#include <lib/cpumask.h>
 
 #include "intf.h"
 #include "types.h"
