+++ scheds/rust/scx_wd40/src/bpf/deadline.bpf.c
@@ -9,7 +9,7 @@
 #include <scx/ravg_impl.bpf.h>
 #include <lib/sdt_task.h>
 
-#include "cpumask.h"
+#include <lib/cpumask.h>
 
 #include <scx/bpf_arena_common.h>
 #include <scx/bpf_arena_spin_lock.h>
