+++ scheds/rust/scx_wd40/src/bpf/placement.bpf.c
@@ -8,7 +8,7 @@
 #include <scx/ravg_impl.bpf.h>
 #include <lib/sdt_task.h>
 
-#include "cpumask.h"
+#include <lib/cpumask.h>
 
 #include <scx/bpf_arena_common.h>
 #include <scx/bpf_arena_spin_lock.h>
